feat:class API
